backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.7
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
mysqlclient==2.2.0

# Authentication & Security
pyotp==2.9.0
qrcode==7.4.2
bcrypt==4.0.1

# ML & Data Science
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
joblib==1.3.2

# NLP & Chatbot
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# Communication
aioredis==2.0.1
celery==5.3.4
aiosmtplib==2.0.2

# Utilities
requests==2.31.0
httpx==0.25.2
aiofiles==23.2.1
pillow==10.1.0

# Export & Reporting
reportlab==4.0.7
openpyxl==3.1.2
matplotlib==3.8.2
plotly==5.17.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0